+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 Unreleased
 ----------
 
+* Add BURDI and CRFDI categorizations and scripts to generate them from the UNFCCC DI
+  flexible query API.
+
 
 0.5.0 (2021-03-23)
 ------------------